COCO-15 + COCO-44 : Finalized Google Binding
Finalized Google Binding to the Calendar page

Fixed Major flaw in app that caused it to run slow & crash.

Some Major Issues Fixed:

The App now instantiates each page and uses it to navigate through the
app without having to create new ones each time.

The Google Binding is included in the Calendar Page only so the user can
access the app directly.

The bugs for the Calendar Page has been fixed.

Removed the activities and everything related to the "choosing" screen
which was added at the beginning.

Other Minor Issues Fixed.