Added Barometer Unit tests, lacking the acceptance of Xamarin.Form in unit test run
All simple tests cases that have been written are now injected. This
will be used to ensure when Xamarin.Forms is accepted as part of the
Unit test, that things are working. Second attempt to get this to work
has failed, we will be reaching out to Shahriar in attempt to rectify
this.

Holding on the more complex tests due to the lack of ability to know if
they are working in their entirety.